+++ R/interactions.R
@@ -1,54 +1,3 @@
-jqui_interaction <- function(ui, interaction,
-                             operation = c("enable", "disable", "destroy", "save", "load"),
-                             options = NULL, switch = NULL) {
-  UseMethod("jqui_interaction")
-}
-
-jqui_interaction.character <- function(ui, interaction,
-                                       operation = c("enable", "disable", "destroy", "save", "load"),
-                                       options = NULL, switch = NULL) {
-  type <- "interaction"
-
-  func <- interaction
-  rm(interaction)
-
-  if (!is.null(switch)) {
-    warning(
-      "The switch parameter is deprecated, ",
-      "please use the operation parameter instead"
-    )
-    operation <- switch(switch, `TRUE` = "enable", `FALSE` = "disable")
-  }
-  rm(switch)
-
-  operation <- match.arg(operation)
-
-  debug <- getOption("shinyjqui.debug")
-
-  sendMsg()
-  return(ui)
-}
-
-jqui_interaction.JS_EVAL <- function(ui, interaction,
-                                     operation = c("enable", "disable", "destroy", "save", "load"),
-                                     options = NULL, switch = NULL) {
-  jqui_interaction.character(ui, interaction, operation, options, switch)
-}
-
-jqui_interaction.shiny.tag <- function(ui, interaction,
-                                       operation = c("enable", "disable", "destroy", "save", "load"),
-                                       options = NULL, switch = NULL) {
-  addInteractJS(ui, interaction, options)
-}
-
-jqui_interaction.shiny.tag.list <- function(ui, interaction,
-                                            operation = c("enable", "disable", "destroy", "save", "load"),
-                                            options = NULL, switch = NULL) {
-  ui[] <- lapply(ui, addInteractJS,
-                       func = interaction, options = options)
-  return(ui)
-}
-
 
 #' Mouse interactions
 #'
@@ -109,8 +58,6 @@ jqui_interaction.shiny.tag.list <- function(ui, interaction,
 #'   Ignored when `operation` is set as `destroy`. This parameter also
 #'   accept a `shiny` option that controls the shiny input value returned from
 #'   the element. See Details.
-#' @param selector,tag,switch Deprecated, just keep for backward compatibility.
-#'   Please use `ui` and `operation` parameters instead.
 #'
 #' @return The same object passed in the `ui` parameter
 #'
@@ -118,152 +65,43 @@ jqui_interaction.shiny.tag.list <- function(ui, interaction,
 #' @name Interactions
 NULL
 
-
-#' @rdname Interactions
-#' @export
-jqui_draggabled <- function(tag, options = NULL) {
-  .Deprecated(
-    msg = "jqui_draggabled() will be deprecated, please use jqui_draggable() instead."
-  )
-  addInteractJS(tag, "draggable", options)
-}
-
-#' @rdname Interactions
-#' @export
-jqui_droppabled <- function(tag, options = NULL) {
-  .Deprecated(
-    msg = "jqui_droppabled() will be deprecated, please use jqui_droppable() instead."
-  )
-  addInteractJS(tag, "droppable", options)
-}
-
-#' @rdname Interactions
-#' @export
-jqui_resizabled <- function(tag, options = NULL) {
-  .Deprecated(
-    msg = "jqui_resizabled() will be deprecated, please use jqui_resizable() instead."
-  )
-  addInteractJS(tag, "resizable", options)
-}
-
-#' @rdname Interactions
-#' @export
-jqui_selectabled <- function(tag, options = NULL) {
-  .Deprecated(
-    msg = "jqui_selectabled() will be deprecated, please use jqui_selectable() instead."
-  )
-  addInteractJS(tag, "selectable", options)
-}
-
-#' @rdname Interactions
-#' @export
-jqui_sortabled <- function(tag, options = NULL) {
-  .Deprecated(
-    msg = "jqui_sortabled() will be deprecated, please use jqui_sortable() instead."
-  )
-  addInteractJS(tag, "sortable", options)
-}
-
 #' @rdname Interactions
 #' @export
 jqui_draggable <- function(ui,
                           operation = c("enable", "disable", "destroy", "save", "load"),
-                          options = NULL, selector = NULL, switch = NULL) {
-  if(!is.null(selector)) {
-    warning(
-      "The selector argument is deprecated, ",
-      "please use the ui argument instead"
-    )
-    ui <- selector
-  }
-  jqui_interaction(ui, "draggable", operation, options, switch)
+                          options = NULL) {
+  jqui_interaction(ui, "draggable", operation, options)
 }
 
 #' @rdname Interactions
 #' @export
 jqui_droppable <- function(ui,
                           operation = c("enable", "disable", "destroy", "save", "load"),
-                          options = NULL, selector = NULL, switch = NULL) {
-  if(!is.null(selector)) {
-    warning(
-      "The selector argument is deprecated, ",
-      "please use the ui argument instead"
-    )
-    ui <- selector
-  }
-  jqui_interaction(ui, "droppable", operation, options, switch)
+                          options = NULL) {
+  jqui_interaction(ui, "droppable", operation, options)
 }
 
 #' @rdname Interactions
 #' @export
 jqui_resizable <- function(ui,
                           operation = c("enable", "disable", "destroy", "save", "load"),
-                          options = NULL, selector = NULL, switch = NULL) {
-  if(!is.null(selector)) {
-    warning(
-      "The selector argument is deprecated, ",
-      "please use the ui argument instead"
-    )
-    ui <- selector
-  }
-  jqui_interaction(ui, "resizable", operation, options, switch)
+                          options = NULL) {
+  jqui_interaction(ui, "resizable", operation, options)
 }
 
 #' @rdname Interactions
 #' @export
 jqui_selectable <- function(ui,
                           operation = c("enable", "disable", "destroy", "save", "load"),
-                          options = NULL, selector = NULL, switch = NULL) {
-  if(!is.null(selector)) {
-    warning(
-      "The selector argument is deprecated, ",
-      "please use the ui argument instead"
-    )
-    ui <- selector
-  }
-  jqui_interaction(ui, "selectable", operation, options, switch)
+                          options = NULL) {
+  jqui_interaction(ui, "selectable", operation, options)
 }
 
 #' @rdname Interactions
 #' @export
 jqui_sortable <- function(ui,
                           operation = c("enable", "disable", "destroy", "save", "load"),
-                          options = NULL, selector = NULL, switch = NULL) {
-  if(!is.null(selector)) {
-    warning(
-      "The selector argument is deprecated, ",
-      "please use the ui argument instead"
-    )
-    ui <- selector
-  }
-  jqui_interaction(ui, "sortable", operation, options, switch)
+                          options = NULL) {
+  jqui_interaction(ui, "sortable", operation, options)
 }
 
-#' Enable bookmarking state of mouse interactions
-#'
-#' Enable shiny
-#' [bookmarking_state](https://shiny.rstudio.com/articles/bookmarking-state.html)
-#' of mouse interactions. By calling this function in `server`, the elements'
-#' `position`, `size`, `selection state` and `sorting state` changed by mouse
-#' operations can be saved and restored through an URL.
-#'
-#' @export
-jqui_bookmarking <- function() {
-  shiny::onRestored(function(state) {
-    inputs <- state$input
-    for (name in names(inputs)) {
-      if (!grepl("__shinyjquiBookmarkState__", name)) next()
-      info <- strsplit(name, "__")[[1]]
-      selector <- paste0("#", info[1])
-      options <- list(state = inputs[[name]])
-      func <- switch(info[3],
-        "draggable" = jqui_draggable,
-        "droppable" = jqui_droppable,
-        "resizable" = jqui_resizable,
-        "sortable" = jqui_sortable,
-        "selectable" = jqui_selectable
-      )
-      func(ui = selector, operation = "load", options = options)
-    }
-  })
-}