+++ NEWS.md
@@ -1,5 +1,9 @@
 ## shinyjqui 0.3.3.9000
 
+* __Breaking change:__ Removed `jqui_draggabled()`, `jqui_droppabled()`, `jqui_resizabled()`, `jqui_selectabled()`, `jqui_sortabled()` and `includeJqueryUI()` as they have been deprecated for a long time.
+* __Breaking change:__ Removed deprecated `selector`, `tag` and `switch` parameters from interation and effect functions.
+* __Bug fix:__ Effect-specific options are no longer ignored by `jqui_effect()`. (#56)
+
 
 ## shinyjqui 0.3.3